Modularized meryl and merqury analysis and added last v1418
